pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[project]
dependencies = [
    "fastapi>=0.114.0",
    "fastapi-msgspec>=0.1.0",
    "granian[reload]>=1.6.0",
    "hypercorn>=0.17.3",
    "loguru>=0.7.2",
    "scalar-fastapi>=1.0.3",
    "python-dotenv>=1.0.1",
    "fastapi-csrf-protect>=0.3.4"
]
description = "Add your description here"
name = "zania-backend"
readme = "README.md"
requires-python = ">=3.12"
version = "0.1.0"

[project.scripts]
dev = "src.zania_backend:main"

[tool.uv]
dev-dependencies = [
    "coverage>=7.6.1",
    "mypy>=1.11.2",
    "pre-commit>=3.8.0",
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
    "pytest-rich>=0.1.1",
    "ruff>=0.6.4"
]
package = true
